fixed takedamage

+++ catGame/GAME1007_W05_Lab4/player1.h
@@ -24,7 +24,7 @@ public:
 	void setPlrPos(int x, int y);
 	int getPlrSpd();
 	int getMaxSpd();
-	void takeDamage(int howMuch);
+	void takeDamage(double howMuch);
 
 };
